requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
itsdangerous==2.2.0
python-multipart==0.0.9
openai>=2.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.11
